Split the Application module for testing.
The new Application module only contains a functor,
Make, which integrates the rest of the code into
an application, providing a single entry point handle_request
that, given a web request, runs the controller as appropriate.

The Main module contains all required glue to rain the appication
as a CGI or FCGI script.